backend/requirements.txt
absl-py==2.1.0
ahocorapy==1.6.2
annotated-types==0.7.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
autocommand==2.2.2
backports.csv==1.0.7
beautifulsoup4==4.12.3
blinker==1.9.0
blis==0.7.11
bottle==0.12.25
branca==0.8.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.4.1
cheroot==10.0.1
CherryPy==18.10.0
click==8.1.8
cloudpathlib==0.18.1
clr-loader==0.2.6
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
debugpy==1.8.5
decorator==5.1.1
distlib==0.3.9
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.7.0/es_core_news_lg-3.7.0-py3-none-any.whl#sha256=08020b83e0c6da1584e567551a5e0de7b15dc0534eaaee21acc1ce908d1be742
et-xmlfile==1.1.0
executing==2.0.1
feedparser==6.0.11
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
folium==0.19.4
fonttools==4.53.1
future==1.0.0
gast==0.6.0
geocoder==1.38.1
geographiclib==2.0
geopandas==1.0.1
geopy==2.4.1
git-filter-repo==2.45.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.12.1
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jaconv==0.4.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jaraco.text==4.0.0
jedi==0.19.1
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.6.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
lute3==3.9.2
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.4.0
murmurhash==1.0.10
mysql-connector==2.2.9
mysqlclient==2.2.4
namex==0.0.8
natto-py==1.0.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
openepub==0.0.8
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
Pattern==3.6
pdfminer.six==20240706
pillow==10.4.0
platformdirs==3.11.0
ply==3.11
portend==3.2.0
preshed==3.0.9
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.29.0
proxy_tools==0.1.0
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.6.0
Pygments==2.18.0
PyMySQL==1.1.1
pyogrio==0.10.0
pyparsing==3.1.4
pypdf==5.1.0
pyproj==3.7.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
pythonnet==3.0.3
pytz==2024.1
pywebview==5.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.1
ratelim==0.1.6
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rsa==4.9
scipy==1.14.1
selenium==4.24.0
setuptools==73.0.1
sgmllib3k==1.0.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.4.8
stack-data==0.6.3
subtitle-parser==2.0.1
tempora==5.7.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
thinc==8.2.5
tk==0.1.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
UpSetPlot==0.9.0
uritemplate==4.1.1
urllib3==2.3.0
vennfig==0.2.3
virtualenv==20.27.1
waitress==2.1.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.2.1
xmltodict==0.14.2
xyzservices==2025.1.0
zc.lockfile==3.0.post1
